competition_service/requirements.txt
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.3.1
tenacity==8.0.1
psycopg2-binary==2.9.3
python-dotenv==0.19.2
Flask-SocketIO==5.3.2
eventlet==0.33.0
dnspython==1.16.0
Werkzeug==1.0.1
SQLAlchemy==1.3.24
requests==2.26.0